% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met2model.STICS.R
\name{met2model.STICS}
\alias{met2model.STICS}
\title{Write STICS met files}
\usage{
met2model.STICS(
  in.path,
  in.prefix,
  outfolder,
  start_date,
  end_date,
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{in.path}{path on disk where CF file lives}

\item{in.prefix}{prefix for each file}

\item{outfolder}{location where model specific output is written.}

\item{start_date}{start date of the simulation}

\item{end_date}{end date of the simulation}

\item{overwrite}{logical: replace output files if they already exist?}

\item{...}{other arguments passed to function}
}
\value{
results
}
\description{
Converts a met CF file to a model specific met file. The input
files are calld <in.path>/<in.prefix>.YYYY.cf
}
\author{
Istem Fer
}
