% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model2netcdf.STICS.R
\name{model2netcdf.STICS}
\alias{model2netcdf.STICS}
\title{Code to convert STICS' output into netCDF format}
\usage{
model2netcdf.STICS(
  outdir,
  sitelat,
  sitelon,
  start_date,
  end_date,
  overwrite = FALSE
)
}
\arguments{
\item{outdir}{Location of model output}

\item{sitelat}{Latitude of the site}

\item{sitelon}{Longitude of the site}

\item{start_date}{Start time of the simulation}

\item{end_date}{End time of the simulation}

\item{overwrite}{Whether or not to overwrite existing output files}
}
\description{
Convert STICS output into the NACP Intercomparison format (ALMA using netCDF)
}
\author{
Istem Fer
}
