% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.config.STICS.R
\name{pecan2stics}
\alias{pecan2stics}
\title{Function to translate pecan param names and units to stics names and units.}
\usage{
pecan2stics(trait.values)
}
\arguments{
\item{trait.values}{trait.values, list}
}
\value{
translated list
}
\description{
Function to translate pecan param names and units to stics names and units.
}
\author{
Quentin Bell
}
