% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.config.STICS.R
\name{write.config.STICS}
\alias{write.config.STICS}
\title{Write STICS configuration files}
\usage{
write.config.STICS(defaults, trait.values, settings, run.id)
}
\arguments{
\item{defaults}{list of defaults to process}

\item{trait.values}{vector of samples for a given trait}

\item{settings}{list of settings from pecan settings file}

\item{run.id}{id of run}
}
\value{
configuration file for STICS for given run
}
\description{
Writes STICS configurations.
}
\details{
Requires a pft xml object, a list of trait values for a single model run,
and the name of the file to create
}
\author{
Istem Fer
}
